requirements.txt
fastapi
uvicorn
sqlmodel
jinja2
python-multipart
python-dotenv
slowapi
